Load .x subsystem first (bugfix)
If we load the templating system first, we get into situations where the templating system adds a directx_x property on the entity, _and_ that leaves it invalidated in the system (which will get picked up on in the next run loop). When the directx system then runs it's on_document_load it finds the directx_x property already there and loads the .x, only to load it again once the invalidation comes through.